AERApi/requirements.txt
psycopg2-binary
Django==3.2.13
djangorestframework==3.13.1
django-filter